requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
yt-dlp>=2023.3.4
psutil>=5.9.0
google-api-python-client>=2.55.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2